refactor: combine filter drawers for projects and news feed pages
